Open Meeting in group notification message after meeting creation

+++ src/main/java/com/razykrashka/bot/stage/meeting/creation/sbs/start/StartNewMeetingCreationSBSStage.java
@@ -16,8 +16,10 @@ public class StartNewMeetingCreationSBSStage extends BaseMeetingCreationSBSStage
     @Override
     public boolean processCallBackQuery() {
         LoadingThreadV2 loadingThread = startLoadingThread(true);
-        meetingRepository.findByCreationStatusEqualsInProgress(updateHelper.getUser().getId())
-                .ifPresent(m -> meetingRepository.delete(m));
+        if (updateHelper.isCallBackDataContains("fromScratch")) {
+            meetingRepository.findByCreationStatusEqualsInProgress(updateHelper.getUser().getId())
+                    .ifPresent(m -> meetingRepository.delete(m));
+        }
         super.setActiveNextStage(DateMeetingCreationSBSStage.class);
         joinToThread(loadingThread);
         razykrashkaBot.getContext().getBean(DateMeetingCreationSBSStage.class).handleRequest();
